Restored Client and Server game loops after pull
